requirements.txt
streamlit>=1.30
pandas>=2.0
google-genai>=1.0
python-dotenv>=1.0
pytest>=7.0
